+++ src/components/common/Modal/hooks/useModal.ts
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 
-import { ModalContext } from './ModalContext';
+import { ModalContext } from '../ModalContext';
 
 const useModal = () => {
   const { isModalOpen, openModal, closeModal } = useContext(ModalContext);
