+++ src/components/common/Modal/hooks/useModalAnimation.ts
